requirements.txt
flask
sqlalchemy
flask-sqlalchemy
python-magic
